feat: move spring to modal
