+++ packages/base-modal/src/index.ts
@@ -1,3 +1,4 @@
 export * from './Component';
 export * from './utils';
 export * from './helpers/lockScroll';
+export { type SpringOptions } from './components/animation-wrapper';
